pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-memory-server"
version = "0.1.0"
description = "Contextra: Agent-Driven MCP Memory Server"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Contextra Team", email = "team@contextra.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "weaviate-client>=4.4.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
    "chromadb>=0.4.0",  # Lightweight alternative to Weaviate
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]
lightweight = [
    # For lightweight deployments without Weaviate/PostgreSQL
    "chromadb>=0.4.0",
    "sqlite>=3.0.0",
]

[project.scripts]
mcp-memory-server = "mcp_memory_server.server:main"

[project.urls]
Homepage = "https://github.com/contextra/mcp-memory-server"
Repository = "https://github.com/contextra/mcp-memory-server"
Documentation = "https://docs.contextra.dev"
"Bug Reports" = "https://github.com/contextra/mcp-memory-server/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src/mcp_memory_server",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
    "-v"
]
asyncio_mode = "auto"
